+++ crates/miden-objects/src/transaction/chain_mmr.rs
@@ -135,10 +135,7 @@ impl ChainMmr {
 impl Serializable for ChainMmr {
     fn write_into<W: miden_crypto::utils::ByteWriter>(&self, target: &mut W) {
         self.mmr.write_into(target);
-        self.blocks.len().write_into(target);
-        for block in self.blocks.values() {
-            block.write_into(target);
-        }
+        self.blocks.write_into(target);
     }
 }
 
@@ -147,12 +144,7 @@ impl Deserializable for ChainMmr {
         source: &mut R,
     ) -> Result<Self, miden_crypto::utils::DeserializationError> {
         let mmr = PartialMmr::read_from(source)?;
-        let block_count = usize::read_from(source)?;
-        let mut blocks = BTreeMap::new();
-        for _ in 0..block_count {
-            let block = BlockHeader::read_from(source)?;
-            blocks.insert(block.block_num(), block);
-        }
+        let blocks = BTreeMap::<BlockNumber, BlockHeader>::read_from(source)?;
         Ok(Self { mmr, blocks })
     }
 }