+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Made NoteFile and AccountFile more consistent (#1133).
 - [BREAKING] Implement most block constraints in `ProposedBlock` (#1123).
 - [BREAKING] Added the `get_block_timestamp` procedure to the `miden` library (#1138).
+- Added serialization for `ProposedBatch`, `BatchId`, `BatchNoteTree` and `ProvenBatch` (#1140).
 
 ## 0.7.2 (2025-01-28) - `miden-objects` crate only